+++ CHANGES.md
@@ -47,6 +47,9 @@ profile. This started with version 0.26.0.
 - Fix `[@ocamlformat "disable"]` inside `class type` constructs. (#2525, @EmileTrotignon)
 - Fix the formatting of the `in` keyword when `[@ocamlformat disable]` is attached to a let-binding (#2242, @EmileTrotignon)
 
+### Changes
+- The location of attributes for structure items is now tracked and preserved. (#2247, @EmileTrotignon)
+
 ## 0.26.1 (2023-09-15)
 
 ### Changed